refactor: change the store to be scalable
