+++ packages/vlossom/src/composables/color-scheme-composable.ts
@@ -5,7 +5,9 @@ import type { Ref } from 'vue';
 import type { ColorScheme, VsComponent } from '@/declaration';
 
 export function useColorScheme(component: VsComponent, colorScheme: Ref<ColorScheme | undefined>) {
-    const computedColorScheme = computed(() => colorScheme.value || store.getGlobalColorScheme(component) || 'default');
+    const computedColorScheme = computed(
+        () => colorScheme.value || store.option.getGlobalColorScheme(component) || 'default',
+    );
 
     return {
         computedColorScheme,
